mimxrt: Use extmod/modnetwork and extmod/modusocket.
This follows the generalization of extmod.
